refactor:barrel exported context related hooks and provider
