+++ src/context/index.js
@@ -1,0 +1,18 @@
+import { useAddress, AddressProvider } from "./address-context";
+import { useFilter, FilterProvider } from "./filter-context";
+import { useCart, CartProvider } from "./cart-context";
+import { WishlistProvider, useWishlist } from "./wishlist-context";
+import { useAuth, AuthProvider } from "./auth-context";
+
+export {
+  useAddress,
+  AddressProvider,
+  useFilter,
+  FilterProvider,
+  useCart,
+  CartProvider,
+  WishlistProvider,
+  useWishlist,
+  useAuth,
+  AuthProvider,
+};
