+++ src/pages/cart/CartSummary.jsx
@@ -1,7 +1,7 @@
 import { useLocation, useNavigate } from "react-router-dom";
 import classes from "./cart.module.css";
 import logo from "../../assets/ecomm-logo.png";
-import { useCart } from "../../context/cart-context";
+import { useCart } from "../../context";
 import InfoAlert from "../../components/Alerts/Info/InfoAlert";
 import { useToggle } from "../../hooks/useToggle";
 import { useState } from "react";
