+++ src/pages/order summary/OrderSummary.jsx
@@ -1,7 +1,6 @@
 import classes from "./orderSummary.module.css";
 import CartSummary from "../cart/CartSummary";
-import { useCart } from "../../context/cart-context";
-import { useAddress } from "../../context/address-context";
+import { useCart,useAddress } from "../../context";
 import { useToggle } from "../../hooks/useToggle";
 import MiniAddressCard from "../../components/mini address card/MiniAddressCard";
 import MiniProductCard from "../../components/mini product card/MiniProductCard";
