+++ src/pages/products/Products.jsx
@@ -7,13 +7,9 @@ import {
   getMaxPrice,
   getCategoryWise,
 } from "../../utils/util";
-
 import classes from "./products.module.css";
-import { useFilter } from "../../context/filter-context";
+import { useFilter,useAuth,useCart,useWishlist } from "../../context";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "../../context/auth-context";
-import { useCart } from "../../context/cart-context";
-import { useWishlist } from "../../context/wishlist-context";
 import SuccessAlert from "../../components/Alerts/Success/SuccessAlert";
 import InfoAlert from "../../components/Alerts/Info/InfoAlert";
 import { useState } from "react";
