+++ src/pages/single product/SingleProduct.jsx
@@ -3,9 +3,7 @@ import { useNavigate, useParams } from "react-router-dom";
 import classes from "./singleProduct.module.css";
 import axios from "axios";
 import Rating from "../../components/rating/Rating";
-import { useWishlist } from "../../context/wishlist-context";
-import { useAuth } from "../../context/auth-context";
-import { useCart } from "../../context/cart-context";
+import { useWishlist,useAuth,useCart } from "../../context";
 import SuccessAlert from "../../components/Alerts/Success/SuccessAlert";
 import InfoAlert from "../../components/Alerts/Info/InfoAlert";
 const SingleProduct = () => {
